simplify add reward modal
